wizards first version
