+++ src/domain/hanke/edit/Form1.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
+import { Button } from 'hds-react';
 
 const Form1: React.FC<any> = (props) => {
   return (
     <div className="form1">
       <h2>Hankkeen alue</h2>
-      <button type="button" onClick={(e) => props.parentCallback(0)}>
+      <Button type="button" onClick={(e) => props.parentCallback(0)}>
         back
-      </button>
-      <button type="button" onClick={(e) => props.parentCallback(2)}>
+      </Button>
+      <Button type="button" onClick={(e) => props.parentCallback(2)}>
         next
-      </button>
+      </Button>
     </div>
   );
 };
