+++ src/domain/hanke/edit/Form2.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
+import { Button } from 'hds-react';
 
 const Form2: React.FC<any> = (props) => {
   return (
     <div className="form2">
       <h2>Hankkeen yhteystiedot</h2>
-      <button type="button" onClick={(e) => props.parentCallback(1)}>
+      <Button type="button" onClick={(e) => props.parentCallback(1)}>
         back
-      </button>
-      <button type="button" onClick={(e) => props.parentCallback(3)}>
+      </Button>
+      <Button type="button" onClick={(e) => props.parentCallback(3)}>
         next
-      </button>
+      </Button>
     </div>
   );
 };
