+++ src/domain/hanke/edit/Form3.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
+import { Button } from 'hds-react';
 
 const Form3: React.FC<any> = (props) => {
   return (
     <div className="form3">
       <h2>Työmaan tiedot</h2>
-      <button type="button" onClick={(e) => props.parentCallback(2)}>
+      <Button type="button" onClick={(e) => props.parentCallback(2)}>
         back
-      </button>
-      <button type="button" onClick={(e) => props.parentCallback(4)}>
+      </Button>
+      <Button type="button" onClick={(e) => props.parentCallback(4)}>
         next
-      </button>
+      </Button>
     </div>
   );
 };
